debt - delete `apiCommands` (fix #110583)
