+++ src/vs/workbench/services/request/browser/requestService.ts
@@ -10,6 +10,8 @@ import { ILogService } from 'vs/platform/log/common/log';
 import { RequestChannelClient } from 'vs/platform/request/common/requestIpc';
 import { IRemoteAgentService, IRemoteAgentConnection } from 'vs/workbench/services/remote/common/remoteAgentService';
 import { RequestService } from 'vs/platform/request/browser/requestService';
+import { ServicesAccessor } from 'vs/editor/browser/editorExtensions';
+import { CommandsRegistry } from 'vs/platform/commands/common/commands';
 
 export class BrowserRequestService extends RequestService {
 
@@ -42,3 +44,15 @@ export class BrowserRequestService extends RequestService {
 		return connection.withChannel('request', channel => RequestChannelClient.request(channel, options, token));
 	}
 }
+
+// --- Internal commands to help authentication for extensions
+
+CommandsRegistry.registerCommand('_workbench.fetchJSON', async function (accessor: ServicesAccessor, url: string, method: string) {
+	const result = await fetch(url, { method, headers: { Accept: 'application/json' } });
+
+	if (result.ok) {
+		return result.json();
+	} else {
+		throw new Error(result.statusText);
+	}
+});
